02 18 2025
I set up the problem quite close to a possible final query.

I enjoyed the process of setting this up

I also got to use the book!

I nestled it as a sort of wrist rest but it is too bulky.  Bad idea.

I may need to figure out a place to get a book holder.