02 25 2025
I have made a breakthrough.  The breakthrough is that the WHERE clause will feed the correct values for the SELECT statement to SUM and divide by two.

I am stuck otherwise.

I am unsure how to proceed with the WHERE clause.  How to  structure it to choose the correct values.