03 25 2025
I am very excited because I am learning how the INDEX works in TSQL.

I am excited but I am afraid that there might not be a way access the Table Scan.  I can be assured that TSQL is using the table scan when I have not or there is not an index access.

so, but how do I tell it to count and the n stop half way after done scanning once?