04 03 2025
Amazing time.

I settled on imagining Example 8.1 after having typed it.

I am really enjoying how the IF statement is used along with the block inside of it.

The block inside of it being the Boolean expression.  The block makes it into a Boolean expression, I think.  Otherwise, would the Boolean expression just be the lines of code in lines 154 - 157.

I may run it next time.