04 06 2025
I have decided to go back and tackle the median problem.

I think with the procedural language IF statement.  I will get it.

I was trying to figure out how to solve for knowing/checking whether a value is odd or even.

I caved and searched on Google only to find an ingenius and straightforward answer.

(column_name % 2) <> 0

This checks if a value is odd.

the (column_name % 2) looks for what the remainder of dividing a value by 2.

I need to grasp this one better.