04 07 2025
ok,

I am certainly closer.

I have however hit a wall in coming up with the Boolean expression.

should it

the ELSE.

I think the Boolean expression should check for 1 or 0.

If 1, then run the (a + b) / 2

If 0, then the result set is just that.  the value that is in the middle of a ranked list of values.

I need to remember to work on the ranking.