04 18 2025
well, I am happy to continue to get lost in the sauce.

I should be able to craft a better approach to getting the median of LAT_N after these two or three more examples.

I am happy to eventually be able to solve this.

I have learned so much today even.  Declaring variables.

And like using TOP to reach an index position but here the index position is used to reach a result set.

All very weird application.