04 20 2025
I am happy to have been able to work on this problem again.

I found a new understanding of what the median means.

It is a very interesting aggregate value thing.

My understanding of SQL shifted along as well.

TOP is already, might already be the median.

I just have to phrase it correctly.