04 22 2025
Happy to report progress!

I am very close, or closer than before to one of the two possibilities.

1. an odd number of rows in a field of numeric items organized ascendingly

or

2. an even number of them.

I think I am close to the even number of rows.

I feel like this could also be a wrong direction.

But I did learn how to make that subquery in lines 935 - 937.