05 04 2025
Really awkward about the fact that I worked on getting long division to see what the modulus was getting.

And to check if there was a remainder pattern between odd or even numbers.

Turns out when divided by 2.  the remainder of an even number is 0 while on the other hand.  the remainder of an odd number when divided by two is 1.

I wanted to check if there was something else like i don't know but I am happy to have this one figured out.