05 06 2025
I thought I had thought of a solid solution to this part of the problem where I am at.  That I need to SUM the two median values that come out of having an even number of values to get the median off of.

But I find myself going back to the OVER clause?

I am unsure but I need to see if I can use the TOP clause in

I am now reading a post where I can use the over clause with the TOP

like TOP (COLUMN) OVER ( PARTITIOBY ETC)