06 10 2025
I am working on learning next time what the ROW_NUMBER() function is.

I tried to do something more direct in query from lines 1849 to 1857 but it didn't work.

Again that determinism vs not-determinist stuff is something I need to flesh out.

But I don't see a way to learn it.  Or a way to grasp it right now except to just run into it again in future problems.