06 23 2025
Ok,

I ended up watching part of a video and have gotten such an interesting answer.

But I need to confirm it because right now it isn't giving me that I consider to  be the median.

It might just be my wording of the query.

I might just need to understand PERCENTILE_CONT ( ) better.